# Pre calc speed up