This branch is a bunch of testing to improve the speed and reduce complexity of hexdump2.  Tested against develop branch (in ipython):
```ipython
d=bytes(range(256))*16
from hexdump2 import hexdump as hd
%timeit _=hd(d,result="return")
6.51 ms ± 46.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
Tested with PR:
```ipython
d=bytes(range(256))*16
from hexdump2 import hexdump as hd
%timeit _=hd(d,result="return")
3.15 ms ± 79.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```